mealScale/Backend/FastAPI-Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]
sqlalchemy
python-dotenv
psycopg2-binary
pydantic-settings
google-auth
python-jose[cryptography]
email-validator
requests
openai
python-multipart